# apps/servienta/openapi.yaml
# The engine's HTTP contract (R11). Ships with the delivery; breaking changes
# bump `info.version`, which GET /api/v1/version reports as `contract`.
openapi: 3.1.0
info:
  title: Servienta engine control API
  version: 0.2.0
paths:
  /api/v1/version:
    get:
      summary: Contract version
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [contract]
                properties: { contract: { type: string } }
  /api/v1/endpoints:
    get:
      summary: Live address of every engine service (R7)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
  /api/v1/license:
    get:
      summary: Current license status and licensed stands (R12, D15)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [mode, stands]
                properties:
                  mode: { type: string, enum: [free, licensed] }
                  stands: { type: array, items: { type: string } }
                  customer: { type: string }
                  expires_at: { type: integer }
                  error: { type: string }
    put:
      summary: Store an issued license and restart to apply it (D20). The body is persisted as-is; the signature is verified at the startup that applies it (R12) — this endpoint grants no signing power (D10).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: The license file exactly as issued }
      responses:
        "202": { description: License stored; the engine restarts to apply it }
        "400": { description: Body is not valid JSON }
    delete:
      summary: Remove the stored license and restart into Free mode (D20)
      responses:
        "202": { description: License removed; the engine restarts }
  /api/v1/reload:
    post:
      summary: Restart in place to re-read the stored license. The engine validates the license only at startup (R12).
      responses:
        "202": { description: Reload accepted; the engine restarts }
  /api/v1/reset:
    post:
      summary: Instance-wide reset to a known state (R5, D3)
      responses:
        "204": { description: Reset complete }
  /api/v1/runs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    put:
      summary: Declare a run and claim its source addresses (R4, D4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sources]
              properties:
                sources:
                  type: array
                  minItems: 1
                  items: { type: string }
      responses:
        "201": { description: Declared }
        "409": { description: Source already claimed, or run already declared }
    delete:
      summary: Release a run's claims (messages stay)
      responses:
        "204": { description: Released }
        "404": { description: Run is not declared }
  /api/v1/responses/{service}:
    put:
      summary: Control a request-response service's reply for the run (R8)
      parameters:
        - { name: service, in: path, required: true, schema: { type: string, enum: [radius, tacacs, dns, ntp] } }
      requestBody:
        required: true
        description: >
          Free-form control the service interprets. radius/tacacs:
          {outcome: accept|reject|pass|fail, reason}. dns:
          {outcome: record|nxdomain|servfail|delay, address, delay_ms}.
          ntp: {outcome: normal|offset|stratum|refuse, offset_ms, stratum}.
          An empty object restores the default (successful reply).
        content:
          application/json:
            schema: { type: object }
      responses:
        "204": { description: Response control set }
  /api/v1/send/{service}:
    post:
      summary: Send a protocol message to a receiving application (R13)
      parameters:
        - { name: service, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        description: >
          {target: "host:port", ...payload}. The target is supplied on every
          request (D18). syslog: {transport, message}. dns: {qname}. radius:
          {secret, username, password}. tacacs: {secret, username}. ntp: {}.
          snmp-traps: {community, oid, value}. kafka: {topic, key, value}.
          ipfix: {src, dst}.
        content:
          application/json:
            schema: { type: object, required: [target], properties: { target: { type: string } } }
      responses:
        "200": { description: "For request-response: the application's reply. For one-way: a confirmation." }
        "404": { description: No licensed sender for this service }
        "502": { description: Send failed (target unreachable, protocol error) }
  /api/v1/received/{service}:
    get:
      summary: Recorded messages in arrival order (R4)
      parameters:
        - { name: service, in: path, required: true, schema: { type: string } }
        - { name: run, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [ts, source, content]
                  properties:
                    ts: { type: string, format: date-time }
                    source: { type: string }
                    run: { type: string }
                    content: { type: object }
        "404": { description: "Unknown receiver (N5) or undeclared run (R4) — never an empty 200" }